requirements.txt
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-tinymce==2.8.0
gunicorn==20.0.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
django-tinymce4-widget==4.0.0
pytz==2019.3
six==1.11.0
whitenoise==5.0.1
Pillow==7.1.1

